pyproject.toml
[project]
name = "typefully-mcp-server"
version = "0.1.0"
description = "MCP server for Typefully API integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
typefully-mcp = "typefully_mcp_server.server:main" 